Convert to Angular SSR (#6)
* dump these changes

* Change to be same setup as new apps

* many refactors...

* try to update the static component which looks a lot better than fullpage...

* Updates to component static and integrating it into page

* updates to the static gospel

* fix up static page

* Update to fix the static site.

* Update to have seo on all pages and fetch ids

* Updates to support ssr in effects.

* Update the static component to be the same as the dynamic one...

* Update blog detail pages

* clean up app routes..

* remove

* change.

* serve option

---------

+++ apps/app/src/app/shared/component/settings-widget.component.ts
@@ -1,5 +1,14 @@
-import { CommonModule } from "@angular/common";
-import { ChangeDetectionStrategy, Component, inject, OnDestroy, OnInit } from "@angular/core";
+import { CommonModule, isPlatformBrowser } from "@angular/common";
+import {
+  afterNextRender,
+  ChangeDetectionStrategy,
+  Component,
+  Inject,
+  inject,
+  OnDestroy,
+  OnInit,
+  PLATFORM_ID,
+} from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { MatButtonModule } from "@angular/material/button";
 import { MatButtonToggleModule } from "@angular/material/button-toggle";
@@ -49,23 +58,27 @@ export class SettingsWidgetComponent implements OnInit, OnDestroy {
     theme: "light",
   };
 
-  constructor() {
-    const storedSettings = localStorage.getItem("appSettings");
-    if (storedSettings) {
-      this.settings = JSON.parse(storedSettings) as AppSettings;
-      this.themeService.setTheme(this.settings.theme);
-    }
+  constructor(@Inject(PLATFORM_ID) private platformId: object) {
+    afterNextRender(() => {
+      const storedSettings = localStorage.getItem("appSettings");
+      if (storedSettings) {
+        this.settings = JSON.parse(storedSettings) as AppSettings;
+        this.themeService.setTheme(this.settings.theme);
+      }
+    });
   }
 
   ngOnInit(): void {}
 
   ngOnDestroy(): void {
+    if (!isPlatformBrowser(this.platformId)) return;
+
     localStorage.clear();
   }
 
   save() {
     const jsonSettings = JSON.stringify({ theme: this.settings.theme });
-    localStorage.setItem("appSettings", jsonSettings);
+    if (isPlatformBrowser(this.platformId)) localStorage.setItem("appSettings", jsonSettings);
 
     this.themeService.setTheme(this.settings.theme);
   }
